Boost: drop the dev-only Overview mock toggle, keep the dummy history fixture
The localStorage `jetpack_boost_overview_mock` toggle was scaffolding for
design review of PR 2. Now that the design has settled, the speed-score,
performance-history, and modules-state hooks each go straight to their
real data sources. The only synthetic piece that stays is
`buildDummyHistory()` — production-required: it powers the blurred curve
under the upgrade Notice that free-tier visitors see on the Historical
performance card.
